+++ jkube-kit/common/src/main/java/org/eclipse/jkube/kit/common/util/SpringBootConfiguration.java
@@ -31,6 +31,8 @@ public class SpringBootConfiguration {
   private String serverKeystore;
   private String managementKeystore;
   private String servletPath;
+  private boolean serverSslEnabled;
+  private boolean managementSslEnabled;
   private String serverContextPath;
   private String managementContextPath;
   private String actuatorBasePath;
@@ -54,27 +56,40 @@ public class SpringBootConfiguration {
       .orElse(1);
     final SpringBootConfiguration.SpringBootConfigurationBuilder configBuilder = SpringBootConfiguration.builder();
     // Spring Boot 1 and common properties
+    final String serverKeystore = properties.getProperty("server.ssl.key-store");
+    final String managementKeystore = properties.getProperty("management.ssl.key-store");
+
     configBuilder
       .managementPort(Optional.ofNullable(properties.getProperty("management.port")).map(Integer::parseInt).orElse(null))
       .serverPort(Integer.parseInt(properties.getProperty("server.port", DEFAULT_SERVER_PORT)))
-      .serverKeystore(properties.getProperty("server.ssl.key-store"))
+      .serverKeystore(serverKeystore)
       .managementHealthProbesEnabled(Boolean.parseBoolean(properties.getProperty("management.health.probes.enabled")))
-      .managementKeystore(properties.getProperty("management.ssl.key-store"))
+      .managementKeystore(managementKeystore)
       .servletPath(properties.getProperty("server.servlet-path"))
       .serverContextPath(properties.getProperty("server.context-path"))
       .managementContextPath(properties.getProperty("management.context-path"))
       .actuatorBasePath("")
       .actuatorDefaultBasePath("")
-      .webFluxBasePath(properties.getProperty("spring.webflux.base-path"));
+      .webFluxBasePath(properties.getProperty("spring.webflux.base-path"))
+      .serverSslEnabled(Optional.ofNullable(properties.getProperty("server.ssl.enabled"))
+        .map(Boolean::parseBoolean)
+        .orElse(serverKeystore != null))
+      .managementSslEnabled(Optional.ofNullable(properties.getProperty("management.ssl.enabled"))
+        .map(Boolean::parseBoolean)
+        .orElse(managementKeystore != null));
     if (majorVersion > 1) {
+      final String managementServerKeystore = properties.getProperty("management.server.ssl.key-store");
       configBuilder
         .managementPort(Optional.ofNullable(properties.getProperty("management.server.port")).map(Integer::parseInt).orElse(null))
-        .managementKeystore(properties.getProperty("management.server.ssl.key-store"))
+        .managementKeystore(managementServerKeystore)
         .servletPath(properties.getProperty("server.servlet.path"))
         .serverContextPath(properties.getProperty("server.servlet.context-path"))
         .managementContextPath(properties.getProperty("management.server.servlet.context-path"))
         .actuatorBasePath(properties.getProperty("management.endpoints.web.base-path"))
-        .actuatorDefaultBasePath("/actuator");
+        .actuatorDefaultBasePath("/actuator")
+        .managementSslEnabled(Optional.ofNullable(properties.getProperty("management.server.ssl.enabled"))
+          .map(Boolean::parseBoolean)
+          .orElse(managementServerKeystore != null));
     }
     if (majorVersion == 3) {
       configBuilder